Finished the final boss

+++ EECS494-P3/Assets/Scripts/HasHealth.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class HasHealth : MonoBehaviour {
-    [SerializeField] protected int maxHealth = 3;
-    protected int health;
+    [SerializeField] protected float maxHealth = 3;
+    protected float health;
 
     private void Awake() {
         Debug.Log("Setting Health to " + maxHealth);
@@ -16,11 +16,11 @@ public class HasHealth : MonoBehaviour {
         health += healthDelta;
     }
 
-    public int GetHealth() {
+    public float GetHealth() {
         return health;
     }
 
-    public int GetMaxHealth() {
+    public float GetMaxHealth() {
         return maxHealth;
     }
 }
